pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "legal_red_line"
version = "1.0.0"
description = "AI Legal Assistant OpenEnv"
dependencies = [
    "openenv-core",
    "pydantic",
    "openai",
    "fastapi",
    "uvicorn"
]

[project.scripts]
server = "server.app:main"
